Supporting unknown AKMs for Wi-Fi connection
If an AP advertise the AKMs which are not defined in
the Wi-Fi framework, those APs will be seen as open network.
So added an overlay configuration item for OEMs to configure
the unknown AKM suite specifier to known AKM suite specifier.
If configured, while parsing the scan result of an AP, Wi-Fi
framework check if the unknown AKM is configured in the overlay
and set the Scan result key management scheme to the mapped
known AKM suite specifier. This approach helps the connect command
to go to wpa_supplicant with a known security type. Then it is upto
the vendor implementation to use the correct AKM for connection.

Updated-Overlayable: TRUE
Bug: 303375329

Test: atest InformationElementUtilTest
Test: Manual - Connect-Disconnect STA-AP connection with WPA2/WPA3
security

Change-Id: I5a0dbcc1714dea1da2c64de309dfa04dd80edee8
Merged-In: I5a0dbcc1714dea1da2c64de309dfa04dd80edee8