wifi: Avoid updating country code when wifi connect/disconnect
Considering that the country code may be pending when CMM is not ready.
The current design is that we check and update country code when any cmm
mode changed. But it is non-necessary if there is no country code is pending
to change.

The major impact is happening on the non-modem device. (for example: watch).
Because those devices observe country code from wifi scan result (802.11d).
Thus we will see country code set to default when cmm mode is changed
(for example: connect/disconnect) and update immediately when receive scan result.

for device with telephoney, the impact is minor since it always update the same
country code (telephony country code).

Bug: 314363189
Test: atest -c FrameworksWifiTests
Change-Id: I1cd8f12f5bef39363dd5848c358e73d505d1f802