Set domain suffix match unconditionally
In case of TOFU connection, the domain suffix match is not set
in the enterprise configuration when subject alternate name is
present in the certificate. But there is nothing wrong in setting
both the domain suffix match and subject alternate name. When supplicant
validate the certificate information, it first checks if the
subject alternate name is present in the enterprise configuration,
if yes then it compares with the subject alternate name present in the
certificate. If subject alternate name is not present, then supplicant
uses domain suffix match for matching. Setting domain suffix match
unconditionally helps to fix the connection failures where subject alternate
name is not set in wifi enterprise configuration due to large size(>
256) of the subject alternate name set in the certificate.

Bug: 317822769
Test: atest WifiConfigManagerTest
Test: atest InsecureEapNetworkHandlerTest
Test: Manual - Connect/Disconnect to WPA3 enterprise AP via TOFU
connection flow

Change-Id: I41125ee31a1dfc341029ff77e698d6032df99cab