Trigger password error for status code of CHALLENGE_FAILURE
As per IEEE 80211-2020 specification section 12.4.7.6, an SAE
confirm message that was not successfully verified is indicated with
a status code of 15 (CHALLENGE_FAILURE). A device fails to verify the
SAE confirm message when the password entered is wrong. So treat
association rejection with a status code of 15 as password error.

Bug: 325331771
Test: atest SupplicantStaIfaceHalAidlImplTest
Test: atest SupplicantStaIfaceHalHidlImplTest
Test: Manual - Enter wrong password and tried to connect to WPA3 AP
      and observed from the logs that AP fails the authentication and sends
      status code=15.
Change-Id: Iaacf6e3487eb523833428ac8c95e5018b2e7b67e