wifi: Fix incorrect default soft ap configuration
The default softap configuration is generated when api
getSoftApConfiguration is called and no persist config exist.
But it means that it may not have chip concurrency information
when generating which was introduced by ag/25575894.

Revert the behavior to check overlay configuration only when
generating default config. But monitoring the chip info updated
event to reset it back to default band to cover those device
which didn't support bridged mode case.

Bug: 327959460
Bug: 328114345
Test: atest -c Frameworks
Test: Manual check, the softap config will be reset when wifi is on.
Change-Id: I691a46fc833e8eae19e2ccd5c975f44719e4e86b