wifi: Support transition disable value in QR code parsing.
When QR code Uri indicates "transition disable", it generates
WPA3 + WPA2 (disabled) network to make sure client won't use WPA2
to connect the access-point. (To avoid lowing security).
It will still show "saved network" in scan list but it won't
be connected. It allows user to know this is disabled by access-point
which same as design when we got "TDI" from access-point.

Bug: 174585999
Test: atest -c FrameworksWifiApiTests
Change-Id: I5e564abb211f4a98a8af4f4a904a6097f20f3cdc