Do oneshot RSSI poll sooner when screen off
Currently when the screen is off, we still do a one time RSSI poll to
update stats after connect. However, this is being done only after L3
connected state which is too late and leaves WifiInfo RSSI be -127 for
the gap which could confuse callers who access this information.

Move this one time poll to L2 connected entering so that external scorer
will get the correct RSSI.

Bug: 330211614
Bug: 329009740
Test: atest com.android.server.wifi
Test: manually verified this works

Change-Id: I85414026ef976441e74c304eb915f317ee0b2c67