Set p2p interface mac address in HIDL implementation
Added the changes to set the p2p interface mac address
in the connected p2p client device information. This is
for the applications to retrive the IP address of the
connected P2P client on a device running Android V system
image and HIDL based HAL implementation.

Bug: 331343436
Test: atest SupplicantP2pIfaceCallbackAidlImplTest
Test: atest SupplicantP2pIfaceCallbackHidlImplTest
Test: Manual - Establish P2P connection and verified from the logs that
both EAPOL IP address allocation and DHCP server are able to set the IP
address.

Change-Id: Ia07885c9604194dfe441695992df60ad53e33503