Drop stale CMD_INTERFACE_DESTROYED message
When there is a country change to a country which support
only 2.4GHz band, the SoftAp teardown the bridged interface
(ap_br_wlan2) and bring up the AP in single Ap mode(wlan2).
The teardown and bring up operation runs in the same context.
The teardown operation triggers the interface destroyed event
for bridged interface. This results in sending the AP disabling
state to UI and on receiving the state change, Wi-Fi hotspot
settings turns off SoftAp. Framework should detect that it is
a just a interface change and avoid sending the state change to
UI. So added a check for AP interface name and the interface name
on which the interface destroyed event message is received.
if they don't match, drop the message.

Bug: 332397333
Test: Manual - Connect STA to AP in channel 149. Turn ON
               SoftAp. SoftAp come up in bridge mode.
	       Set country code to Japan using wifi
	       shell command. Confirmed from the logs that
	       CMD_INTERFACE_DESTROYED is not processed when
	       this event is received after single AP interface is
	       enabled. Also verified that Wi-Fi Hotspot UI status
	       remains in ON state.
Change-Id: I6e7a479b48d58cc949b6eeccfb01f4ce6325da66