Fix supplicant stuck trying to connect with fallback SSID
When SSID translation is on, we attempt to connect to a fallback SSID if
the first SSID we try is not found. However, if the first SSID was
actually the correct one but the first attempt failed, then supplicant
will be stuck trying to connect with the fallback SSID.

To fix this, alternate between the first SSID and the fallback on each
network not found notification.

Bug: 311117929
Test: atest FrameworksWifiTests
Change-Id: I2e2a400df75fec4062bd32eb48565da57df0aa12