Additional cleanup during wifi disable
The intention is to always do certain cleanup in the
WifiConnectivityManager when the user toggles wifi. However, today the
cleanup may not happen in some cases when IMS over wifi is on and
the user toggles wifi relatively fast.

To fix the problem, we need to add a more explicit trigger for the
cleanup when wifi is toggled.

Bug: 332601371
Test: atest com.android.server.wifi

Change-Id: Ideaac4aea42e86003138e7467db736ea7989f37d