Connection failure when SubjectAltName contains dNSName
The TOFU connection failure happens when the subject
alternative name in the certificate contains the
domain name system(for eg: DNS:wi-fi.org). In this case
the domain_suffix_match is matched against the dNSName
in the SubjectAltName. A CL (Change-ID -
I41125ee31a1dfc341029ff77e698d6032df99cab ) was added recently
which sets the domain suffix match as common name in the subject
unconditionally. This was added to fix another connection failure
caused by a large SubjectAltName field. This change caused the
certificate verification failure where the common name is matched
against the dNSName. So to fix both the issues increased the maximum
allowed SubjectAltName length from 256 bytes to 1024 and
reverted the changes added in Change-Id - I41125ee31a1dfc341029ff77e698d6032df99cab .
With this change, if there is a subject alternative name in the
certificate, only the subject alternative name will be set in the
enterprise configuration. If there is no subject alternative name, then
the common name will be set as the domain suffix match in the
configuration.

Bug: 335099099
Bug: 315485525
Test: atest WifiConfigManagerTest
Test: Manual - TOFU connection
Change-Id: Ifd2be24c347aaa593bb3d3787e9a67dccd599569