Modified the check for Maximum Number of AKMs
The Wi-Fi Alliance has introduced the WPA3 security update for Wi-Fi 7, which
mandates cross-AKM (Authenticated Key Management) roaming between three AKMs
(AKM: 24(SAE-EXT-KEY), AKM:8(SAE) and AKM:2(PSK)). If the station supports
AKM 24(SAE-EXT-KEY), it is recommended to enable WPA3 SAE auto-upgrade offload,
provided that the driver indicates that the maximum number of AKM suites allowed in
connection requests is three or more.

If the driver doesn't indicate the maximum number of AKM suites,
kernel defaults to 2. So the current check for maximum number of AKM
to 2 doesn't guarantee that the driver supports cross-akm roaming.
So modified the check to 3 or more.

Bug: 336732588
Test: atest ClientModeImplTest
Test: Manual - Tested and confirmed from the logs that cross-akm feature
is getting enabled.

Change-Id: I60d1d923655f24922841a8ffa9c6b514558be70b