Fix startTetheredHotspot failure during SAP teardown
Tethering may restart hotspot in case of a downstream prefix conflict.
The new call to startTetheredHotspot is triggered when the iface goes
down, but this may fail if the SAP state hasn't changed to DISABLED yet.
To fix this, set the state to DISABLED before tearing down the iface to
ensure startTetheredHotspot can be called immediately again.

Bug: 335478233
Test: atest SoftApManagerTest
Change-Id: Id31a6443091819dc576b67398f57196720d972ff