Set PMF required when connecting to WPA3 network
When user explicitly configure a network to WPA3 Personal or
WPA3 Enterprise (Via Add network option in settings), the PMF required
bit should be set in the association request RSN capabilities. So don't
turn off the PMF required bit in security params when the AP is
configured in transition network and when the local secuirty params in
not added by auto upgrade.

Bug: 338519250
Test: Manual - Configured a WPA3-TM network in settings -> Add network ->
with security type WPA3 and confirmed from the sniffer logs that PMF
required bit is set in Association request RSN capabilities.

Change-Id: Iba047b5b1161a7df1cb15752967de5da3a01c622