P2P: Generate IPV6 address based on EUI64
Specify in p2p client provisioning configuration to generate
IPv6 link local only address from the interface's
MAC address using the modified EUI-64 format. This
change is done due to the following reasons,
1. Kernel default configuration to generate the
IPV6 address (addr_gen_mode) is based on EUI64.
On the P2P GO side, network stack/wifi framework has no
way to overwrite this configuration. So the IPV6 address of GO
is derived based on the P2P GO interface MAC address. Whereas
on the client side the network stack overwrite the configuration
to generate the IPV6 address based on the secret from stable_secret
configuration. So the IPV6 address is generared based on different modes
on P2P client and GO. So on devices which uses the same static interface
for P2P device interface and group interface will be using the last
updated configuration from network stack for deriving the IPV6 address.
So when a device changes the role from P2P Client to P2P GO, the IPV6
address will be derived based on the secret. So in order to have the
same IPV6 address generation mode reliably, wifi p2p framework needs to set the mode
on P2P client side.
2. The IPV6 provisioning mode - GROUP_CLIENT_IP_PROVISIONING_MODE_IPV6_LINK_LOCAL
on the P2P client side is designed based on the reasoning that the P2P
GO always generate the IPV6 address based on EUI64.

Bug: 334762957
Test: Manual - Establish P2P connection and verified using
ifconfig <interface name> on both P2P GO and P2P CLIENT side
that IPV6 address is based on EUI-64 format

Change-Id: I4e5f6467c5e588505f98d9f4e0c42417daac1f90