wifi: Add a new overlay to indicate device supports wep allowed control
The WEP control toggle UI may not exist for all android devices
(example: watch) but the settings will be backup/restore cross those devices. To make sure the user has a way to control WEP network (allowed / disallowed) when WEP network is disallowed from restoration.
Add a new overlay to indicate the device supports WEP allowed control
(default is true) and use WEP allowed settings only when the control is
supported. Otherwise the user may face that it can't connect to WEP
network but it didn't have a user interface to control it.

Bug: 328817046
Test: atest -c FrameworksWifiTests
Change-Id: I5759e2463c22be3dc7f0201c6f4240d92287c81c