Explicitly specify the jarjar rule for generated flag classes
The change explicitly specify the jarjar fule for generated flag
classes. The reason is the jarjar tool can transform the literals in the
code based on the rule. If the rule is too general, jarjar will
transform the string constant in the generated code. The constant is
used for accessing the storage, and for FlaggedApi usage. So the
constant should not be transformed.

Test: presubmit
Bug: 358435763
Change-Id: I05c08b1c033f869116a87e12c4c41302cfca68be