Prevent visible background users from making changes to WIFI.
No behavior changes on a non-Automotive device.
On an Automotive device, setWifiEnabled() will ignore calls from an
invalid user, which is neither the system user, the current user, nor
a profile of the current user. The following methods will throw a
SecurityException for calls from an invalid user.
- startTetheredHotspot
- startTetheredHotspotRequest
- stopSoftAp

Bug: 360488316
Flag: EXEMPT bugfix
Test: atest WifiServiceImplTest
Test: Build gcar_md_x86_64 (with a local patch to enable Wifi controls)
      Start emulator
      adb logcat AndroidRuntime:e *:s
      adb shell setprop fw.visible_bg_users true
      Start a pasenger user on a passenger screen
      Go to Settings and enable/diable WiFi

Change-Id: I391711c359174c1f3a8ef19edec82c29d7ff11ab